# WIP: Use FillArrays and LazyArrays to reduce memory footprint