--- a/REQUIRE
+++ b/REQUIRE
@@ -1,3 +1,5 @@
 julia 0.7
 MathProgBase 0.7 0.8
 DataStructures
+LazyArrays
+FillArrays