--- a/src/atoms/affine/diag.jl
+++ b/src/atoms/affine/diag.jl
@@ -83,7 +83,7 @@ function conic_form!(x::DiagAtom, unique_conic_forms::UniqueConicForms=UniqueCon
             sz_diag = Base.min(num_rows + k, num_cols)
         end
 
-        select_diag = spzeros(sz_diag, length(x.children[1]))
+        select_diag = zeros(sz_diag, length(x.children[1]))
         for i in 1:sz_diag
             select_diag[i, start_index] = 1
             start_index += num_rows + 1