--- a/src/atoms/exp_cone/log.jl
+++ b/src/atoms/exp_cone/log.jl
@@ -49,7 +49,7 @@ function conic_form!(e::LogAtom, unique_conic_forms::UniqueConicForms=UniqueConi
     if !has_conic_form(unique_conic_forms, e)
         # log(z) \geq x  <=>    (x,ones(),z) \in ExpCone
         z = e.children[1]
-        y = Constant(ones(size(z)))
+        y = Constant(Ones(size(z)))
         x = Variable(size(z))
         objective = conic_form!(x, unique_conic_forms)
         conic_form!(ExpConstraint(x, y, z), unique_conic_forms)