--- a/src/atoms/second_order_cone/qol_elementwise.jl
+++ b/src/atoms/second_order_cone/qol_elementwise.jl
@@ -49,9 +49,9 @@ end
 
 qol_elementwise(x::AbstractExpr, y::AbstractExpr) = QolElemAtom(x, y)
 
-broadcasted(::typeof(^),x::AbstractExpr,k::Int) = k==2 ? QolElemAtom(x, Constant(ones(x.size[1], x.size[2]))) : error("raising variables to powers other than 2 is not implemented")
+broadcasted(::typeof(^),x::AbstractExpr,k::Int) = k==2 ? QolElemAtom(x, Constant(Ones(x.size[1], x.size[2]))) : error("raising variables to powers other than 2 is not implemented")
 
-invpos(x::AbstractExpr) = QolElemAtom(Constant(ones(x.size[1], x.size[2])), x)
+invpos(x::AbstractExpr) = QolElemAtom(Constant(Ones(x.size[1], x.size[2])), x)
 broadcasted(::typeof(/), x::Value, y::AbstractExpr) = DotMultiplyAtom(Constant(x), invpos(y))
 /(x::Value, y::AbstractExpr) = size(y) == (1,1) ? MultiplyAtom(Constant(x), invpos(y)) : error("cannot divide by a variable of size $(size(y))")
 sumsquares(x::AbstractExpr) = square(norm2(x))
@@ -60,6 +60,6 @@ function square(x::AbstractExpr)
     if sign(x) == ComplexSign()
         error("Square of complex number is not DCP. Did you mean square_modulus?")
     else
-        QolElemAtom(x, Constant(ones(x.size[1], x.size[2])))
+        QolElemAtom(x, Constant(Ones(x.size[1], x.size[2])))
     end
 end