@ararslan 2018-12-22
Right now this is just "throw laziness at the wall and see what sticks"; I'm mostly making sure that this approach doesn't full-on break anything. Doing long-running computations and monitoring `top` shows drastically decreased memory usage on this branch, but I don't have individual benchmarks at the moment. I'll do that more formally as I fine-tune this.