@ararslan 2019-01-02
Unfortunately, for the simplest problem:
```julia
x = Variable(1)
p = minimize(2.0 * x, [x >= 2, x <= 4])
```
this PR provides a pretty modest benefit in terms of memory use but a regression in terms of execution time.

Master:
```
julia> @benchmark Convex.conic_problem($p)
BenchmarkTools.Trial: 
  memory estimate:  38.77 KiB
  allocs estimate:  703
  --------------
  minimum time:     34.333 μs (0.00% GC)
  median time:      42.053 μs (0.00% GC)
  mean time:        138.870 μs (68.14% GC)
  maximum time:     504.959 ms (99.98% GC)
  --------------
  samples:          10000
  evals/sample:     1
```
PR:
```
julia> @benchmark Convex.conic_problem($p)
BenchmarkTools.Trial: 
  memory estimate:  30.17 KiB
  allocs estimate:  613
  --------------
  minimum time:     66.901 μs (0.00% GC)
  median time:      76.468 μs (0.00% GC)
  mean time:        107.625 μs (25.10% GC)
  maximum time:     24.867 ms (99.13% GC)
  --------------
  samples:          10000
  evals/sample:     1
```
Pretty disappointing. I'll keep at this.