@ararslan 2019-01-03
Observation: LazyArrays has some facilities to act as a computation graph in some scenarios, e.g. with `Mul` representing a multiplication. There is a direct analog here with Convex's own `MultiplyAtom`. I wonder if there's a way to more fundamentally combine these two.