@odow 2022-02-23
Closing as stale. 

Moving more to MOI (e.g., https://github.com/jump-dev/Convex.jl/pull/393) seems like a better way of dealing with the memory issues.